requirements.txt
beautifulsoup4
lxml
tqdm
pandas
pydantic
PyYAML
Requests
# certifi
